+++ Android/app/src/main/java/com/dergoogler/core/NativeUtils.java
@@ -1,0 +1,16 @@
+package com.dergoogler.core;
+
+import android.webkit.JavascriptInterface;
+
+public class NativeUtils {
+
+    @JavascriptInterface
+    public void runtimeException(String message) {
+        throw new RuntimeException(message);
+    }
+
+    @JavascriptInterface
+    public void exception(String message) throws Exception {
+        throw new Exception(message);
+    }
+}