+++ Website/src/components/Markdown/rules.ts
@@ -1,11 +1,11 @@
 export const rules = [
   // Headers
-  [/={6}\s?([^\n]+)/g, "<divider><h6>$1</h6></divider>"],
-  [/={5}\s?([^\n]+)/g, "<divider><h5>$1</h5></divider>"],
-  [/={4}\s?([^\n]+)/g, "<divider><h4>$1</h4></divider>"],
-  [/={3}\s?([^\n]+)/g, "<divider><h3>$1</h3></divider>"],
-  [/={2}\s?([^\n]+)/g, "<divider><h2>$1</h2></divider>"],
-  [/={1}\s?([^\n]+)/g, "<divider><h1>$1</h1></divider>"],
+  [/#{6}\s?(([^\n]+)?\s)#{6}(?:\n|$)/g, "<divider><h6>$2</h6></divider>"],
+  [/#{5}\s?(([^\n]+)?\s)#{5}(?:\n|$)/g, "<divider><h5>$2</h5></divider>"],
+  [/#{4}\s?(([^\n]+)?\s)#{4}(?:\n|$)/g, "<divider><h4>$2</h4></divider>"],
+  [/#{3}\s?(([^\n]+)?\s)#{3}(?:\n|$)/g, "<divider><h3>$2</h3></divider>"],
+  [/#{2}\s?(([^\n]+)?\s)#{2}(?:\n|$)/g, "<divider><h2>$2</h2></divider>"],
+  [/#{1}\s?(([^\n]+)?\s)#{1}(?:\n|$)/g, "<divider><h1>$2</h1></divider>"],
 
   // Marks
   [/{{checkmark}}/g, "<checkmark/>"],